Fix the communication of reader versions.
It has previously been assumed that reader version on the managing
machine and the batch machine were the same. This is of course not always
true, and moreover one machine may not even be able to get a version,
such as with the newly setup running of ISI. Reader versions are now
reported to s3 by each job, and aggregated by the managing machine.