Use _conn and cursor to get raw statements.
This ensures that the agent insert occurs entirely within the same
transaction as the other copies. Otherwise, if the previous copy of
statements had not been committed, they would not show up, and you could
not do everything as part of one transaction, possibly leading to
statements entering the database without agents.