Refactor supplement method of preassembly.
Note that I implement the same change to using queries by hash
as in the previous commit, except that the iteration over the
old (existing) pa statements, which is loaded the old-fashioned
way because we do not already have hashes in hand and the list
of hashes will be very large. It occurs to me that it was
probably this stage that made me first switch away from hash-based
queries.

I also added a type constraint when appropriate, which I realized
I had missed in previous commits. Whoops.