Avoid extraineous complexes.
When getting agent pairs, complex results show up in
all their permutations. Make it so they only show up
once, and preferably such that they will almost always
show up with the agents that were queried (because
there will in general be more examples of such).

This is done by not returning agent JSONs for which
all represented hashes are complexes that have already
been represented. This means the limits and offsets no
longer directly map to those sent to the database,
which was long maintained for simplicity. However,
the machinery already existed to supply a custom offset
for the next page in the response, which is now once
again utilized in this case.

To allow for continued filtering across paging, the
complex hashes that were already displayed are sent as
part of the response, and it is recommended that paging
queries return this set in the next query as part of
the class initialization.